Added CDIF (Cadence Data Interchange Format) encoding/decoding and used it for parsing of events / other cdif payloadds

+++ src/main/kotlin/org/onflow/sdk/AsyncFlowAccessApi.kt
@@ -1,6 +1,5 @@
 package org.onflow.sdk
 
-import java.math.BigInteger
 import java.util.concurrent.CompletableFuture
 
 interface AsyncFlowAccessApi {
@@ -11,13 +10,13 @@ interface AsyncFlowAccessApi {
 
     fun getBlockHeaderById(id: FlowId): CompletableFuture<FlowBlockHeader?>
 
-    fun getBlockHeaderByHeight(height: BigInteger): CompletableFuture<FlowBlockHeader?>
+    fun getBlockHeaderByHeight(height: Long): CompletableFuture<FlowBlockHeader?>
 
     fun getLatestBlock(sealed: Boolean = true): CompletableFuture<FlowBlock>
 
     fun getBlockById(id: FlowId): CompletableFuture<FlowBlock?>
 
-    fun getBlockByHeight(height: BigInteger): CompletableFuture<FlowBlock?>
+    fun getBlockByHeight(height: Long): CompletableFuture<FlowBlock?>
 
     fun getCollectionById(id: FlowId): CompletableFuture<FlowCollection?>
 
@@ -35,15 +34,15 @@ interface AsyncFlowAccessApi {
 
     fun getAccountAtLatestBlock(addresss: FlowAddress): CompletableFuture<FLowAccount?>
 
-    fun getAccountByBlockHeight(addresss: FlowAddress, height: BigInteger): CompletableFuture<FLowAccount?>
+    fun getAccountByBlockHeight(addresss: FlowAddress, height: Long): CompletableFuture<FLowAccount?>
 
     fun executeScriptAtLatestBlock(script: FlowScript): CompletableFuture<FlowScriptResponse>
 
     fun executeScriptAtBlockId(script: FlowScript, blockId: FlowId): CompletableFuture<FlowScriptResponse>
 
-    fun executeScriptAtBlockHeight(script: FlowScript, height: BigInteger): CompletableFuture<FlowScriptResponse>
+    fun executeScriptAtBlockHeight(script: FlowScript, height: Long): CompletableFuture<FlowScriptResponse>
 
-    fun getEventsForHeightRange(type: String, range: ClosedRange<BigInteger>): CompletableFuture<List<FlowEventResult>>
+    fun getEventsForHeightRange(type: String, range: ClosedRange<Long>): CompletableFuture<List<FlowEventResult>>
 
     fun getEventsForBlockIds(type: String, ids: Set<FlowId>): CompletableFuture<List<FlowEventResult>>
